+++ package.json
@@ -18,7 +18,7 @@
   "dependencies": {
     "@koa/cors": "^3.1.0",
     "@koa/router": "^10.0.0",
-    "bull": "^4.8.4",
+    "bullmq": "^1.86.4",
     "bytes": "^3.1.2",
     "debug": "^4.3.3",
     "dotenv-safe": "^8.2.0",
